UT/EVT refactoring
