+++ fsw/crypto_tests/CMakeLists.txt
@@ -22,3 +22,7 @@ add_test(NAME Process_Security
 add_test(NAME UT_TC_APPLY
          COMMAND ${PROJECT_BINARY_DIR}/bin/ut_tc_apply 
          WORKING_DIRECTORY ${PROJECT_TEST_DIR})
+
+add_test(NAME ET_VALIDATION
+         COMMAND ${PROJECT_BINARY_DIR}/bin/et_validation 
+         WORKING_DIRECTORY ${PROJECT_TEST_DIR})
