+++ fsw/crypto_tests/data/encryption_test_ping.txt
@@ -1,1 +1,1 @@
-20030415001880d2ca0008197f0b0031000039c53128
+20030015001880d2ca0008197f0b0031000039c53128
